Fix invalid format specifier in PartialWebhookState.__getattr__
The message for the AttributeError raised by the __getattr__ method of
the PartialWebhookState class is formatted using `str.format`. However,
the placeholder contained a stray ":" before the !r specifier. This
caused a ValueError("Invalid format specifier") to be raised whenever
this method was called in lieu of the AttributeError that is intended
to be raised.